requirements.txt
pandas
SQLAlchemy
psycopg2-binary
python-dotenv

fastapi
uvicorn